Fix Unittests build fails when LTO is disabled Related issue: #774
JerryScript-DCO-1.0-Signed-off-by: SaeHie Park saehie.park@samsung.com
