+++ tests/unit/test-api.cpp
@@ -737,38 +737,3 @@ main (void)
 
   return 0;
 }
-
-/**
- * Provide log message to filestream implementation for the engine.
- */
-int jerry_port_logmsg (FILE* stream, const char* format, ...)
-{
-  va_list args;
-  int count;
-  va_start (args, format);
-  count = vfprintf (stream, format, args);
-  va_end (args);
-  return count;
-}
-
-/**
- * Provide error message to console implementation for the engine.
- */
-int jerry_port_errormsg (const char* format, ...)
-{
-  va_list args;
-  int count;
-  va_start (args, format);
-  count = vfprintf (stderr, format, args);
-  va_end (args);
-  return count;
-}
-
-
-/**
- * Provide output character to console implementation for the engine.
- */
-int jerry_port_putchar (int c)
-{
-  return putchar (c);
-}
